+++ pkg/local_object_storage/shard/delete.go
@@ -74,6 +74,9 @@ func (s *Shard) Delete(prm DeletePrm) (DeleteRes, error) {
 
 	s.decObjectCounterBy(physical, res.RawObjectsRemoved())
 	s.decObjectCounterBy(logical, res.AvailableObjectsRemoved())
+	for i := range prm.addr {
+		s.addToContainerSize(prm.addr[i].Container().EncodeToString(), -int64(res.RemovedObjectSizes()[i]))
+	}
 
 	for i := range prm.addr {
 		var delPrm common.DeletePrm