+++ pkg/local_object_storage/shard/metrics_test.go
@@ -16,26 +16,27 @@ import (
 )
 
 type metricsStore struct {
-	s map[string]uint64
+	objCounters map[string]uint64
+	cnrSize     map[string]int64
 }
 
 func (m metricsStore) SetShardID(_ string) {}
 
 func (m metricsStore) SetObjectCounter(objectType string, v uint64) {
-	m.s[objectType] = v
+	m.objCounters[objectType] = v
 }
 
 func (m metricsStore) AddToObjectCounter(objectType string, delta int) {
 	switch {
 	case delta > 0:
-		m.s[objectType] += uint64(delta)
+		m.objCounters[objectType] += uint64(delta)
 	case delta < 0:
 		uDelta := uint64(-delta)
 
-		if m.s[objectType] >= uDelta {
-			m.s[objectType] -= uDelta
+		if m.objCounters[objectType] >= uDelta {
+			m.objCounters[objectType] -= uDelta
 		} else {
-			m.s[objectType] = 0
+			m.objCounters[objectType] = 0
 		}
 	case delta == 0:
 		return
@@ -43,11 +44,15 @@ func (m metricsStore) AddToObjectCounter(objectType string, delta int) {
 }
 
 func (m metricsStore) IncObjectCounter(objectType string) {
-	m.s[objectType] += 1
+	m.objCounters[objectType] += 1
 }
 
 func (m metricsStore) DecObjectCounter(objectType string) {
 	m.AddToObjectCounter(objectType, -1)
+}
+
+func (m metricsStore) AddToContainerSize(cnr string, size int64) {
+	m.cnrSize[cnr] += size
 }
 
 const physical = "phy"
@@ -64,9 +69,16 @@ func TestCounters(t *testing.T) {
 	}
 
 	t.Run("defaults", func(t *testing.T) {
-		require.Zero(t, mm.s[physical])
-		require.Zero(t, mm.s[logical])
+		require.Zero(t, mm.objCounters[physical])
+		require.Zero(t, mm.objCounters[logical])
+		require.Empty(t, mm.cnrSize)
 	})
+
+	expectedSizes := make(map[string]int64)
+	for i := range oo {
+		cnr, _ := oo[i].ContainerID()
+		expectedSizes[cnr.EncodeToString()] += int64(oo[i].PayloadSize())
+	}
 
 	t.Run("put", func(t *testing.T) {
 		var prm shard.PutPrm
@@ -78,8 +90,9 @@ func TestCounters(t *testing.T) {
 			require.NoError(t, err)
 		}
 
-		require.Equal(t, uint64(objNumber), mm.s[physical])
-		require.Equal(t, uint64(objNumber), mm.s[logical])
+		require.Equal(t, uint64(objNumber), mm.objCounters[physical])
+		require.Equal(t, uint64(objNumber), mm.objCounters[logical])
+		require.Equal(t, expectedSizes, mm.cnrSize)
 	})
 
 	t.Run("inhume_GC", func(t *testing.T) {
@@ -93,8 +106,9 @@ func TestCounters(t *testing.T) {
 			require.NoError(t, err)
 		}
 
-		require.Equal(t, uint64(objNumber), mm.s[physical])
-		require.Equal(t, uint64(objNumber-inhumedNumber), mm.s[logical])
+		require.Equal(t, uint64(objNumber), mm.objCounters[physical])
+		require.Equal(t, uint64(objNumber-inhumedNumber), mm.objCounters[logical])
+		require.Equal(t, expectedSizes, mm.cnrSize)
 
 		oo = oo[inhumedNumber:]
 	})
@@ -103,8 +117,8 @@ func TestCounters(t *testing.T) {
 		var prm shard.InhumePrm
 		ts := objectcore.AddressOf(generateObject(t))
 
-		phy := mm.s[physical]
-		logic := mm.s[logical]
+		phy := mm.objCounters[physical]
+		logic := mm.objCounters[logical]
 
 		inhumedNumber := int(phy / 4)
 		prm.SetTarget(ts, addrFromObjs(oo[:inhumedNumber])...)
@@ -112,8 +126,9 @@ func TestCounters(t *testing.T) {
 		_, err := sh.Inhume(prm)
 		require.NoError(t, err)
 
-		require.Equal(t, phy, mm.s[physical])
-		require.Equal(t, logic-uint64(inhumedNumber), mm.s[logical])
+		require.Equal(t, phy, mm.objCounters[physical])
+		require.Equal(t, logic-uint64(inhumedNumber), mm.objCounters[logical])
+		require.Equal(t, expectedSizes, mm.cnrSize)
 
 		oo = oo[inhumedNumber:]
 	})
@@ -121,8 +136,8 @@ func TestCounters(t *testing.T) {
 	t.Run("Delete", func(t *testing.T) {
 		var prm shard.DeletePrm
 
-		phy := mm.s[physical]
-		logic := mm.s[logical]
+		phy := mm.objCounters[physical]
+		logic := mm.objCounters[logical]
 
 		deletedNumber := int(phy / 4)
 		prm.SetAddresses(addrFromObjs(oo[:deletedNumber])...)
@@ -130,8 +145,13 @@ func TestCounters(t *testing.T) {
 		_, err := sh.Delete(prm)
 		require.NoError(t, err)
 
-		require.Equal(t, phy-uint64(deletedNumber), mm.s[physical])
-		require.Equal(t, logic-uint64(deletedNumber), mm.s[logical])
+		require.Equal(t, phy-uint64(deletedNumber), mm.objCounters[physical])
+		require.Equal(t, logic-uint64(deletedNumber), mm.objCounters[logical])
+		for i := range oo[:deletedNumber] {
+			cnr, _ := oo[i].ContainerID()
+			expectedSizes[cnr.EncodeToString()] -= int64(oo[i].PayloadSize())
+		}
+		require.Equal(t, expectedSizes, mm.cnrSize)
 	})
 }
 
@@ -148,10 +168,11 @@ func shardWithMetrics(t *testing.T, path string) (*shard.Shard, *metricsStore) {
 	}
 
 	mm := &metricsStore{
-		s: map[string]uint64{
+		objCounters: map[string]uint64{
 			"phy":   0,
 			"logic": 0,
 		},
+		cnrSize: make(map[string]int64),
 	}
 
 	sh := shard.New(